+++ src/rpc/avalanche.cpp
@@ -1874,6 +1874,75 @@ static RPCHelpMan getstakecontendervote() {
     };
 }
 
+static RPCHelpMan finalizetransaction() {
+    return RPCHelpMan{
+        "finalizetransaction",
+        "Force finalize a mempool transaction. No attempt is made to poll for "
+        "this transaction and this could cause the node to disagree with the "
+        "network. This can fail if the transaction to be finalized would "
+        "overflow the block size. Upon success it will be included in the "
+        "block template.\n",
+        {
+            {"txid", RPCArg::Type::STR_HEX, RPCArg::Optional::NO,
+             "The id of the transaction to be finalized."},
+        },
+        RPCResult{RPCResult::Type::ARR,
+                  "finalized_txids",
+                  "The list of the successfully finalized txids if any (it can "
+                  "include ancestors of the target txid).",
+                  {{
+                      RPCResult::Type::STR_HEX,
+                      "txid",
+                      "The finalized transaction id.",
+                  }}},
+        RPCExamples{HelpExampleRpc("finalizetransaction", "<txid>")},
+        [&](const RPCHelpMan &self, const Config &config,
+            const JSONRPCRequest &request) -> UniValue {
+            const NodeContext &node = EnsureAnyNodeContext(request.context);
+            CTxMemPool &mempool = EnsureAnyMemPool(request.context);
+            const ChainstateManager &chainman = EnsureChainman(node);
+
+            const TxId txid(ParseHashV(request.params[0], "txid"));
+
+            LOCK2(cs_main, mempool.cs);
+            auto entry = mempool.GetIter(txid);
+            if (!entry) {
+                throw JSONRPCError(RPC_INVALID_PARAMETER,
+                                   "The transaction is not in the mempool.");
+            }
+
+            const CBlockIndex *tip = chainman.ActiveTip();
+            if (!tip) {
+                throw JSONRPCError(RPC_INTERNAL_ERROR,
+                                   "There is no active chain tip.");
+            }
+
+            UniValue ret(UniValue::VARR);
+
+            std::vector<TxId> finalizedTxids;
+            if (!mempool.setAvalancheFinalized(**entry, chainman.GetConsensus(),
+                                               *tip, finalizedTxids)) {
+                // If the function returned false, the finalizedTxids vector
+                // should not be relied upon
+                return ret;
+            }
+
+            for (TxId &finalizedTxid : finalizedTxids) {
+                ret.push_back(finalizedTxid.ToString());
+
+                // FIXME we might want to remove from the recent rejects as well
+                // if it exists so we don't vote against this tx anymore. For
+                // now this is a private data from the PeerManager and can only
+                // be cleared entirely. Also a rejected transaction is not
+                // expected to be in the mempool in the first place so this is
+                // probably safe.
+            }
+
+            return ret;
+        },
+    };
+}
+
 void RegisterAvalancheRPCCommands(CRPCTable &t) {
     // clang-format off
     static const CRPCCommand commands[] = {
@@ -1902,6 +1971,7 @@ void RegisterAvalancheRPCCommands(CRPCTable &t) {
         { "avalanche",         verifyavalanchedelegation, },
         { "avalanche",         setflakyproof,             },
         { "avalanche",         getflakyproofs,            },
+        { "avalanche",         finalizetransaction,       },
         { "hidden",            getavailabilityscore,      },
         { "hidden",            getstakecontendervote,     },
     };