+++ doc/release-notes.md
@@ -4,4 +4,9 @@ Bitcoin ABC version 0.32.1 is now available from:
 
   <https://download.bitcoinabc.org/0.32.1/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - A new `finalizetransaction` is added to force finalize a transaction locally.
+   This makes it possible for miners to add a mempool transaction to the block
+   template, but it should be used with caution as the node might be out of
+   consensus with the Avalanche stakers. See the RPC documentation for more
+   details.